--- components/log-viewer-webui/server/src/fastify-v2/plugins/app/search/SearchResultsMetadataCollection/typings.ts
@@ -1,33 +1,0 @@
-import {Nullable} from "../../../../../typings/common.js";
-
-
-/**
- * Enum of search-related signals.
- *
- * This includes request and response signals for various search operations and their respective
- * states.
- */
-enum SEARCH_SIGNAL {
-    NONE = "none",
-
-    REQ_CANCELLING = "req-cancelling",
-    REQ_CLEARING = "req-clearing",
-    REQ_QUERYING = "req-querying",
-
-    RESP_DONE = "resp-done",
-    RESP_QUERYING = "resp-querying",
-}
-
-/**
- * MongoDB document for search results metadata. `numTotalResults` is optional
- * since it is only set when the search job is completed.
- */
-interface SearchResultsMetadataDocument {
-    _id: string;
-    errorMsg: Nullable<string>;
-    lastSignal: SEARCH_SIGNAL;
-    numTotalResults?: number;
-}
-
-export type {SearchResultsMetadataDocument};
-export {SEARCH_SIGNAL};
